huge refactor of resources hooks
